Move Widget struct to Widget.h
