+++ src/openrct2/interface/Widget.h
@@ -13,53 +13,155 @@
 
 struct WindowBase;
 
-using WidgetIndex = uint16_t;
-
-constexpr WidgetIndex kWidgetIndexNull = 0xFFFF;
-
-enum class WindowWidgetType : uint8_t
+namespace OpenRCT2
 {
-    Empty = 0,
-    Frame = 1,
-    Resize = 2,
-    ImgBtn = 3,
-    ColourBtn = 6,
-    TrnBtn = 7,
-    Tab = 8,
-    FlatBtn = 9,
-    Button = 10,
-    LabelCentred = 12, // Centred text
-    TableHeader = 13,  // Left-aligned textual button
-    Label = 14,        // Left-aligned text
-    Spinner = 15,
-    DropdownMenu = 16,
-    Viewport = 17,
-    Groupbox = 19,
-    Caption = 20,
-    CloseBox = 21,
-    Scroll = 22,
-    Checkbox = 23,
-    Placeholder = 25,
-    ProgressBar = 29,
-    Custom = 28,
-    TextBox = 27,
-};
+    using WidgetIndex = uint16_t;
+    constexpr WidgetIndex kWidgetIndexNull = 0xFFFF;
 
-constexpr uint8_t kCloseButtonWidth = 10;
+    struct WidgetIdentifier
+    {
+        WindowIdentifier window;
+        WidgetIndex widget_index;
+    };
 
-constexpr int32_t kScrollableRowHeight = 12;
-constexpr uint8_t kListRowHeight = 12;
-constexpr uint8_t kTableCellHeight = 12;
-constexpr uint8_t kButtonFaceHeight = 12;
-constexpr uint8_t kSpinnerHeight = 12;
-constexpr uint8_t kDropdownHeight = 12;
+    enum class WindowWidgetType : uint8_t
+    {
+        Empty = 0,
+        Frame = 1,
+        Resize = 2,
+        ImgBtn = 3,
+        ColourBtn = 6,
+        TrnBtn = 7,
+        Tab = 8,
+        FlatBtn = 9,
+        Button = 10,
+        LabelCentred = 12, // Centred text
+        TableHeader = 13,  // Left-aligned textual button
+        Label = 14,        // Left-aligned text
+        Spinner = 15,
+        DropdownMenu = 16,
+        Viewport = 17,
+        Groupbox = 19,
+        Caption = 20,
+        CloseBox = 21,
+        Scroll = 22,
+        Checkbox = 23,
+        Placeholder = 25,
+        ProgressBar = 29,
+        Custom = 28,
+        TextBox = 27,
+    };
 
-constexpr uint16_t kTextInputSize = 1024;
-constexpr uint16_t kTopToolbarHeight = 27;
+    struct Widget
+    {
+        WindowWidgetType type{};
+        uint8_t colour{};
+        int16_t left{};
+        int16_t right{};
+        int16_t top{};
+        int16_t bottom{};
+        union
+        {
+            uint32_t content;
+            ImageId image{};
+            StringId text;
+            utf8* string;
+        };
+        StringId tooltip{ STR_NONE };
 
-enum
-{
-    SCROLL_HORIZONTAL = (1 << 0),
-    SCROLL_VERTICAL = (1 << 1),
-    SCROLL_BOTH = SCROLL_HORIZONTAL | SCROLL_VERTICAL
-};
+        // New properties
+        WidgetFlags flags{};
+        utf8* sztooltip{};
+
+        int16_t width() const
+        {
+            return right - left;
+        }
+
+        int16_t height() const
+        {
+            return bottom - top;
+        }
+
+        int16_t midX() const
+        {
+            return (left + right) / 2;
+        }
+
+        int16_t midY() const
+        {
+            return (top + bottom) / 2;
+        }
+
+        int16_t textTop() const
+        {
+            if (height() >= 10)
+                return std::max<int32_t>(top, top + (height() / 2) - 5);
+
+            return top - 1;
+        }
+
+        void moveRight(int32_t amount)
+        {
+            left += amount;
+            right += amount;
+        }
+
+        void moveDown(int32_t amount)
+        {
+            top += amount;
+            bottom += amount;
+        }
+
+        void moveTo(ScreenCoordsXY coords)
+        {
+            moveRight(coords.x - left);
+            moveDown(coords.y - top);
+        }
+
+        void moveToX(int16_t x)
+        {
+            moveRight(x - left);
+        }
+
+        void moveToY(int16_t y)
+        {
+            moveDown(y - top);
+        }
+
+        bool IsVisible() const
+        {
+            return !(flags & OpenRCT2::WIDGET_FLAGS::IS_HIDDEN);
+        }
+    };
+
+    using WidgetFlags = uint32_t;
+    namespace WIDGET_FLAGS
+    {
+        const WidgetFlags TEXT_IS_STRING = 1 << 0;
+        const WidgetFlags IS_PRESSED = 1 << 2;
+        const WidgetFlags IS_DISABLED = 1 << 3;
+        const WidgetFlags TOOLTIP_IS_STRING = 1 << 4;
+        const WidgetFlags IS_HIDDEN = 1 << 5;
+        const WidgetFlags IS_HOLDABLE = 1 << 6;
+    } // namespace OpenRCT2::WIDGET_FLAGS
+
+    enum
+    {
+        SCROLL_HORIZONTAL = (1 << 0),
+        SCROLL_VERTICAL = (1 << 1),
+        SCROLL_BOTH = SCROLL_HORIZONTAL | SCROLL_VERTICAL
+    };
+
+    constexpr uint8_t kCloseButtonWidth = 10;
+
+    constexpr int32_t kScrollableRowHeight = 12;
+    constexpr uint8_t kListRowHeight = 12;
+    constexpr uint8_t kTableCellHeight = 12;
+    constexpr uint8_t kButtonFaceHeight = 12;
+    constexpr uint8_t kSpinnerHeight = 12;
+    constexpr uint8_t kDropdownHeight = 12;
+
+    constexpr uint16_t kTextInputSize = 1024;
+    constexpr uint16_t kTopToolbarHeight = 27;
+}
