+++ src/openrct2/interface/Window.h
@@ -52,109 +52,7 @@ struct WindowIdentifier
     rct_windownumber number;
 };
 
-struct WidgetIdentifier
-{
-    WindowIdentifier window;
-    WidgetIndex widget_index;
-};
-
 extern WindowCloseModifier gLastCloseModifier;
-
-using WidgetFlags = uint32_t;
-namespace OpenRCT2::WIDGET_FLAGS
-{
-    const WidgetFlags TEXT_IS_STRING = 1 << 0;
-    const WidgetFlags IS_PRESSED = 1 << 2;
-    const WidgetFlags IS_DISABLED = 1 << 3;
-    const WidgetFlags TOOLTIP_IS_STRING = 1 << 4;
-    const WidgetFlags IS_HIDDEN = 1 << 5;
-    const WidgetFlags IS_HOLDABLE = 1 << 6;
-} // namespace OpenRCT2::WIDGET_FLAGS
-
-enum class WindowWidgetType : uint8_t;
-
-struct Widget
-{
-    WindowWidgetType type{};
-    uint8_t colour{};
-    int16_t left{};
-    int16_t right{};
-    int16_t top{};
-    int16_t bottom{};
-    union
-    {
-        uint32_t content;
-        ImageId image{};
-        StringId text;
-        utf8* string;
-    };
-    StringId tooltip{ STR_NONE };
-
-    // New properties
-    WidgetFlags flags{};
-    utf8* sztooltip{};
-
-    int16_t width() const
-    {
-        return right - left;
-    }
-
-    int16_t height() const
-    {
-        return bottom - top;
-    }
-
-    int16_t midX() const
-    {
-        return (left + right) / 2;
-    }
-
-    int16_t midY() const
-    {
-        return (top + bottom) / 2;
-    }
-
-    int16_t textTop() const
-    {
-        if (height() >= 10)
-            return std::max<int32_t>(top, top + (height() / 2) - 5);
-
-        return top - 1;
-    }
-
-    void moveRight(int32_t amount)
-    {
-        left += amount;
-        right += amount;
-    }
-
-    void moveDown(int32_t amount)
-    {
-        top += amount;
-        bottom += amount;
-    }
-
-    void moveTo(ScreenCoordsXY coords)
-    {
-        moveRight(coords.x - left);
-        moveDown(coords.y - top);
-    }
-
-    void moveToX(int16_t x)
-    {
-        moveRight(x - left);
-    }
-
-    void moveToY(int16_t y)
-    {
-        moveDown(y - top);
-    }
-
-    bool IsVisible() const
-    {
-        return !(flags & OpenRCT2::WIDGET_FLAGS::IS_HIDDEN);
-    }
-};
 
 /**
  * Viewport structure
